HHH-6615 - Refactoring revision number to long

+++ hibernate-envers/src/matrix/java/org/hibernate/envers/test/integration/basic/SingleOperationInTransaction.java
@@ -80,29 +80,29 @@ public class SingleOperationInTransaction extends AbstractEntityTest {
 
     @Test
     public void testRevisionsCounts() {
-        assert Arrays.asList(1, 4, 6).equals(getAuditReader().getRevisions(BasicTestEntity1.class, id1));
+        assert Arrays.asList(1L, 4L, 6L).equals(getAuditReader().getRevisions(BasicTestEntity1.class, id1));
 
-        assert Arrays.asList(2, 5, 7).equals(getAuditReader().getRevisions(BasicTestEntity1.class, id2));
+        assert Arrays.asList(2L, 5L, 7L).equals(getAuditReader().getRevisions(BasicTestEntity1.class, id2));
 
-        assert Arrays.asList(3).equals(getAuditReader().getRevisions(BasicTestEntity1.class, id3));
+        assert Arrays.asList(3L).equals(getAuditReader().getRevisions(BasicTestEntity1.class, id3));
     }
 
     @Test
     public void testRevisionsDates() {
-        for (int i=1; i<7; i++) {
+        for (long i=1L; i<7L; i++) {
             assert getAuditReader().getRevisionDate(i).getTime() <=
-                    getAuditReader().getRevisionDate(i+1).getTime();
+                    getAuditReader().getRevisionDate(i+1L).getTime();
         }
     }
 
     @Test(expected = RevisionDoesNotExistException.class)
     public void testNotExistingRevision() {
-        getAuditReader().getRevisionDate(8);
+        getAuditReader().getRevisionDate(8L);
     }
 
     @Test(expected = IllegalArgumentException.class)
     public void testIllegalRevision() {
-        getAuditReader().getRevisionDate(0);
+        getAuditReader().getRevisionDate(0L);
     }
 
     @Test
@@ -111,13 +111,13 @@ public class SingleOperationInTransaction extends AbstractEntityTest {
         BasicTestEntity1 ver2 = new BasicTestEntity1(id1, "x2", 2);
         BasicTestEntity1 ver3 = new BasicTestEntity1(id1, "x3", 3);
 
-        assert getAuditReader().find(BasicTestEntity1.class, id1, 1).equals(ver1);
-        assert getAuditReader().find(BasicTestEntity1.class, id1, 2).equals(ver1);
-        assert getAuditReader().find(BasicTestEntity1.class, id1, 3).equals(ver1);
-        assert getAuditReader().find(BasicTestEntity1.class, id1, 4).equals(ver2);
-        assert getAuditReader().find(BasicTestEntity1.class, id1, 5).equals(ver2);
-        assert getAuditReader().find(BasicTestEntity1.class, id1, 6).equals(ver3);
-        assert getAuditReader().find(BasicTestEntity1.class, id1, 7).equals(ver3);
+        assert getAuditReader().find(BasicTestEntity1.class, id1, 1L).equals(ver1);
+        assert getAuditReader().find(BasicTestEntity1.class, id1, 2L).equals(ver1);
+        assert getAuditReader().find(BasicTestEntity1.class, id1, 3L).equals(ver1);
+        assert getAuditReader().find(BasicTestEntity1.class, id1, 4L).equals(ver2);
+        assert getAuditReader().find(BasicTestEntity1.class, id1, 5L).equals(ver2);
+        assert getAuditReader().find(BasicTestEntity1.class, id1, 6L).equals(ver3);
+        assert getAuditReader().find(BasicTestEntity1.class, id1, 7L).equals(ver3);
     }
 
     @Test
@@ -126,32 +126,32 @@ public class SingleOperationInTransaction extends AbstractEntityTest {
         BasicTestEntity1 ver2 = new BasicTestEntity1(id2, "y2", 20);
         BasicTestEntity1 ver3 = new BasicTestEntity1(id2, "y3", 21);
 
-        assert getAuditReader().find(BasicTestEntity1.class, id2, 1) == null;
-        assert getAuditReader().find(BasicTestEntity1.class, id2, 2).equals(ver1);
-        assert getAuditReader().find(BasicTestEntity1.class, id2, 3).equals(ver1);
-        assert getAuditReader().find(BasicTestEntity1.class, id2, 4).equals(ver1);
-        assert getAuditReader().find(BasicTestEntity1.class, id2, 5).equals(ver2);
-        assert getAuditReader().find(BasicTestEntity1.class, id2, 6).equals(ver2);
-        assert getAuditReader().find(BasicTestEntity1.class, id2, 7).equals(ver3);
+        assert getAuditReader().find(BasicTestEntity1.class, id2, 1L) == null;
+        assert getAuditReader().find(BasicTestEntity1.class, id2, 2L).equals(ver1);
+        assert getAuditReader().find(BasicTestEntity1.class, id2, 3L).equals(ver1);
+        assert getAuditReader().find(BasicTestEntity1.class, id2, 4L).equals(ver1);
+        assert getAuditReader().find(BasicTestEntity1.class, id2, 5L).equals(ver2);
+        assert getAuditReader().find(BasicTestEntity1.class, id2, 6L).equals(ver2);
+        assert getAuditReader().find(BasicTestEntity1.class, id2, 7L).equals(ver3);
     }
 
     @Test
     public void testHistoryOfId3() {
         BasicTestEntity1 ver1 = new BasicTestEntity1(id3, "z", 30);
 
-        assert getAuditReader().find(BasicTestEntity1.class, id3, 1) == null;
-        assert getAuditReader().find(BasicTestEntity1.class, id3, 2) == null;
-        assert getAuditReader().find(BasicTestEntity1.class, id3, 3).equals(ver1);
-        assert getAuditReader().find(BasicTestEntity1.class, id3, 4).equals(ver1);
-        assert getAuditReader().find(BasicTestEntity1.class, id3, 5).equals(ver1);
-        assert getAuditReader().find(BasicTestEntity1.class, id3, 6).equals(ver1);
-        assert getAuditReader().find(BasicTestEntity1.class, id3, 7).equals(ver1);
+        assert getAuditReader().find(BasicTestEntity1.class, id3, 1L) == null;
+        assert getAuditReader().find(BasicTestEntity1.class, id3, 2L) == null;
+        assert getAuditReader().find(BasicTestEntity1.class, id3, 3L).equals(ver1);
+        assert getAuditReader().find(BasicTestEntity1.class, id3, 4L).equals(ver1);
+        assert getAuditReader().find(BasicTestEntity1.class, id3, 5L).equals(ver1);
+        assert getAuditReader().find(BasicTestEntity1.class, id3, 6L).equals(ver1);
+        assert getAuditReader().find(BasicTestEntity1.class, id3, 7L).equals(ver1);
     }
 
     @Test
     public void testHistoryOfNotExistingEntity() {
-        assert getAuditReader().find(BasicTestEntity1.class, id1+id2+id3, 1) == null;
-        assert getAuditReader().find(BasicTestEntity1.class, id1+id2+id3, 7) == null;
+        assert getAuditReader().find(BasicTestEntity1.class, id1+id2+id3, 1L) == null;
+        assert getAuditReader().find(BasicTestEntity1.class, id1+id2+id3, 7L) == null;
     }
 
     @Test
